+++ receiver/otlpreceiver/internal/logs/otlp_test.go
@@ -70,7 +70,7 @@ func TestExport_ErrorConsumer(t *testing.T) {
 	assert.Equal(t, plogotlp.Response{}, resp)
 }
 
-func makeLogsServiceClient(t *testing.T, lc consumer.Logs) plogotlp.Client {
+func makeLogsServiceClient(t *testing.T, lc consumer.Logs) plogotlp.GRPCClient {
 	addr := otlpReceiverOnGRPCServer(t, lc)
 	cc, err := grpc.Dial(addr.String(), grpc.WithTransportCredentials(insecure.NewCredentials()), grpc.WithBlock())
 	require.NoError(t, err, "Failed to create the TraceServiceClient: %v", err)